+++ geofileops/util/_ogr_util.py
@@ -448,7 +448,7 @@ def vector_translate(
     output_ds = None
     try:
         # Till gdal 3.10 datetime columns can be interpreted wrongly with arrow.
-        if _compat.GDAL_STE_310:
+        if _compat.GDAL_ST_311 and "OGR2OGR_USE_ARROW_API" not in config_options:
             config_options["OGR2OGR_USE_ARROW_API"] = False
 
         # Go!
@@ -546,7 +546,9 @@ def vector_translate(
         if output_ds is None:
             raise RuntimeError("output_ds is None")
 
-        # Sometimes an invalid output file is written, so close and try to reopen it.
+        # Fix invalig files that were written with older versions of GDAL.
+        # More info: https://github.com/geofileops/geofileops/issues/313
+        # if not _compat.GDAL_GTE_38:
         output_ds = None
         _validate_file(
             output_path,